pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["MohammadMahmoodM <mohammadmahmood441@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langchain-core = "^0.3.5"
langchain-community = "^0.3.0"
langchain-openai = "^0.2.0"
python-dotenv = "^1.0.1"
langchain-google-genai = "^2.0.0"
fastapi = "^0.115.0"
faiss-cpu = "^1.8.0.post1"
uvicorn = "^0.30.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
